[Feat] 선택한 식물의 식물일기 리스트로 이동/ 내 식물 수정,삭제 구현

+++ springboot/src/main/java/com/tave_app_1/senapool/user/dto/UserUpdateDto.java
@@ -7,7 +7,7 @@ import javax.persistence.Id;
 @Getter
 public class UserUpdateDto {
     @Id
-    private Integer userPk;
+    private Long userPk;
 
     private String userId;
